refactor: split SmartCheckout.ts into 2 forms for CVC & CC Input
